Refactor tour to abstract away main component
