+++ src/components/tour/constants.js
